+++ footer.php
@@ -1,3 +1,0 @@
-<?php wp_footer(); ?>
-</body>
-</html>