+++ src/app.js
@@ -7,13 +7,16 @@ window.SETTINGS = {
 }
 
 require('./bootstrap')
-require('./vendor/fontawesome/packs/solid.js')
-require('./vendor/fontawesome/fontawesome.js')
 
 import 'babel-polyfill'
 import Vue from 'vue'
 import VueRouter from 'vue-router'
 import VueLazyload from 'vue-lazyload'
+
+import BootstrapVue from 'bootstrap-vue'
+import 'bootstrap/dist/css/bootstrap.css'
+import 'bootstrap-vue/dist/bootstrap-vue.css'
+
 import router from './router'
 // import { auth, database } from './firebase.config.js' - Uncomment if you need firebase
 import App from './App.vue'
@@ -21,6 +24,7 @@ import store from './store'
 import * as types from './store/mutation-types'
 
 Vue.use(VueLazyload)
+Vue.use(BootstrapVue)
 
 router.afterEach((to, from) => {
   // Add a body class specific to the route we're viewing