+++ src/components/Home.vue
@@ -1,24 +1,25 @@
 <template>
-    <div class="container">
-        <div class="jumbotron">
-            <h1 class="display-3">Hello, world!</h1>
-            <p class="lead">This is a simple hero unit, a simple jumbotron-style component for calling extra attention to featured content or information.</p>
-            <hr class="my-4">
-            <p>It uses utility classes for typography and spacing to space content out within the larger container.</p>
-            <p class="lead">
-                <a class="btn btn-primary btn-lg" href="#" role="button">Learn more</a>
-            </p>
-        </div>
+    <b-container class="bv-example-row">
+        <b-jumbotron header="Bootstrap Vue" lead="Bootstrap 4 Components for Vue.js 2" >
+        <p>For more information visit website</p>
+        <b-btn variant="primary" href="https://bootstrap-vue.js.org" target="_blank">More Info</b-btn>
+        </b-jumbotron>
 
-        <div class="row">
-            <div class="col-xs-12 col-md-4">
+        <b-row>
+            <b-col>
                 <recent-posts-widget limit="5">Recent Posts</recent-posts-widget>
-            </div>
-            <div class="col-xs-12 col-md-4">
+            </b-col>
+            <b-col>
                 <pages-widget limit="5">Pages</pages-widget>
-            </div>
-        </div>
-    </div>
+            </b-col>
+            <b-col>
+                <h3>Welcome!</h3>
+                <p>You'll probably want to remove all of this stuff, but hey it at least gives you a preview of Bootstrap-Vue!</p>
+                <p>I would encourage you to avoid using jQuery and instead find the Vue way of doing some things. You'll thank me later!</p>
+                <p>~ <a href="http://evanagee.com/">Evan Agee</a></p>
+            </b-col>
+        </b-row>
+    </b-container>
 </template>
 
 <script>